Enable DTLS protocol negotiation
The new negotiation is as follows:
If the client's X-DTLS-CipherSuite contains the PSK keyword, the
server will reply with "X-DTLS-CipherSuite: PSK" and will enable
DTLS-PSK negotiation on the DTLS channel. The ciphersuite set in the
DTLS channel, must match the one set in TLS one. That, makes the protocol
consistent in security properties (DTLS and TLS channel will match cipher/mac
combinations), and allows the protocol to use new DTLS versions, as well as new
DTLS ciphersuites, as long as they are also used in the TLS channel establishment.

That change still requires to client to pretend it is resuming by setting in the
TLS client hello the session ID provided by the X-DTLS-Session-ID header. That is,
because there is no TLS extension we can use to set an identifier in the client
hello (draft-jay-tls-psk-identity-extension could be used in the future).